requirements.txt
myst_parser
tqdm>=4.47.0
munch>=2.5.0
torch>=1.7.1
opencv_python_headless>=4.1.1.26
requests>=2.22.0
einops>=0.3.0
x_transformers==0.15.0
transformers>=4.18.0
tokenizers==0.12.1
numpy>=1.19.5
Pillow==9.0.0
PyYAML>=5.4.1
pandas>=1.0.0
timm==0.5.4
albumentations>=0.5.2
python-Levenshtein>=0.12.2
torchtext>=0.6.0
imagesize>=1.2.0
scikit-learn
wandb